+++ python/dolfinx_mpc/multipointconstraint.py
@@ -18,15 +18,6 @@ from .dictcondition import create_dictionary_constraint
 from .periodic_condition import create_periodic_condition_topological, create_periodic_condition_geometrical
 
 
-def cpp_dirichletbc(bc):
-    """Unwrap Dirichlet BC objects as cpp objects"""
-    if isinstance(bc, _fem.DirichletBC):
-        return bc._cpp_object
-    elif isinstance(bc, (tuple, list)):
-        return list(map(lambda sub_bc: cpp_dirichletbc(sub_bc), bc))
-    return bc
-
-
 class MultiPointConstraint():
     """
     Multi-point constraint class.
@@ -106,7 +97,7 @@ class MultiPointConstraint():
 
     def create_periodic_constraint_topological(self, meshtag: _mesh.MeshTags, tag: int,
                                                relation: Callable[[numpy.ndarray], numpy.ndarray],
-                                               bcs: list([_fem.DirichletBC]), scale: _PETSc.ScalarType = 1):
+                                               bcs: list([_fem.DirichletBCMetaClass]), scale: _PETSc.ScalarType = 1):
         """
         Create periodic condition for all dofs in MeshTag with given marker:
         u(x_i) = scale * u(relation(x_i))
@@ -133,7 +124,8 @@ class MultiPointConstraint():
     def create_periodic_constraint_geometrical(self, V: _fem.FunctionSpace,
                                                indicator: Callable[[numpy.ndarray], numpy.ndarray],
                                                relation: Callable[[numpy.ndarray], numpy.ndarray],
-                                               bcs: List[_fem.DirichletBC], scale: _PETSc.ScalarType = 1):
+                                               bcs: List[_fem.DirichletBCMetaClass], scale: _PETSc.ScalarType = 1,
+                                               subspace: int = None):
         """
         Create a periodic condition for all degrees of freedom whose physical location satisfies indicator(x)
         u(x_i) = scale * u(relation(x_i)) for all x_i where indicator(x_i) == True
@@ -151,12 +143,12 @@ class MultiPointConstraint():
             Float for scaling bc
         """
         slaves, masters, coeffs, owners, offsets = create_periodic_condition_geometrical(
-            self.V, indicator, relation, bcs, scale)
+            self.V, indicator, relation, bcs, scale, subspace)
         self.add_constraint(self.V, slaves, masters, coeffs, owners, offsets)
 
     def create_slip_constraint(self, facet_marker: tuple([_mesh.MeshTags, int]), v: _fem.Function,
                                sub_space: _fem.FunctionSpace = None, sub_map: numpy.ndarray = numpy.array([]),
-                               bcs: list([_fem.DirichletBC]) = []):
+                               bcs: list([_fem.DirichletBCMetaClass]) = []):
         """
         Create a slip constraint dot(u, v)=0 over the entities defined in a `dolfinx.mesh.MeshTags`
         marked with index i. normal is the normal vector defined as a vector function.
@@ -193,7 +185,7 @@ class MultiPointConstraint():
              W0 = W.sub(0)
              V, V_to_W = W0.collapse(True)
              n = Function(V)
-             bc = _fem.DirichletBC(inlet_velocity, dofs, W0)
+             bc = dirichletbc(inlet_velocity, dofs, W0)
              create_slip_constraint((mt, i), normal, V, V_to_W, bcs=[bc])
         """
         if sub_space is None:
@@ -203,7 +195,7 @@ class MultiPointConstraint():
         mesh_tag, marker = facet_marker
         mpc_data = dolfinx_mpc.cpp.mpc.create_slip_condition(W, mesh_tag, marker, v._cpp_object,
                                                              numpy.asarray(sub_map, dtype=numpy.int32),
-                                                             cpp_dirichletbc(bcs))
+                                                             bcs)
         self.add_constraint_from_mpc_data(self.V, mpc_data=mpc_data)
 
     def create_general_constraint(self, slave_master_dict: Dict[bytes, Dict[bytes, float]],