+++ python/tests/test_mpc_pipeline.py
@@ -44,9 +44,9 @@ def test_pipeline(master_point, get_assemblers):  # noqa: F811
     a = d * g * ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
     rhs = h * ufl.inner(f, v) * ufl.dx
     # Generate reference matrices
-    A_org = fem.assemble_matrix(a)
+    A_org = fem.assemble_matrix(fem.form(a))
     A_org.assemble()
-    L_org = fem.assemble_vector(rhs)
+    L_org = fem.assemble_vector(fem.form(rhs))
     L_org.ghostUpdate(addv=PETSc.InsertMode.ADD_VALUES, mode=PETSc.ScatterMode.REVERSE)
 
     # Create multipoint constraint
@@ -122,9 +122,9 @@ def test_linearproblem(master_point):
     a = d * g * ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
     rhs = h * ufl.inner(f, v) * ufl.dx
     # Generate reference matrices
-    A_org = fem.assemble_matrix(a)
+    A_org = fem.assemble_matrix(fem.form(a))
     A_org.assemble()
-    L_org = fem.assemble_vector(rhs)
+    L_org = fem.assemble_vector(fem.form(rhs))
     L_org.ghostUpdate(addv=PETSc.InsertMode.ADD_VALUES, mode=PETSc.ScatterMode.REVERSE)
 
     # Create multipoint constraint