+++ python/tests/test_vector_assembly.py
@@ -48,7 +48,7 @@ def test_mpc_assembly(master_point, degree, celltype, get_assemblers):  # noqa: 
     b.ghostUpdate(addv=PETSc.InsertMode.ADD_VALUES, mode=PETSc.ScatterMode.REVERSE)
 
     # Reduce system with global matrix K after assembly
-    L_org = fem.assemble_vector(rhs)
+    L_org = fem.assemble_vector(fem.form(rhs))
     L_org.ghostUpdate(addv=PETSc.InsertMode.ADD_VALUES, mode=PETSc.ScatterMode.REVERSE)
     root = 0
     comm = mesh.comm